requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3 
dash-bootstrap-components==0.10.0
dash-daq==0.5.0
dash-renderer==1.4.1
dash-table==4.7.0
pymongo==3.10.1
psycopg2==2.8.5
paho-mqtt==1.5.0
dask==2.1.0
decorator==4.4.0
Flask==1.1.1
Flask-Compress==1.4.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jupyter-core==4.5.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.4
pandas==0.24.2
plotly==4.1.0
pytz==2019.1
six==1.12.0
traitlets==4.3.2
Werkzeug==0.15.4
